network-manager: support a defaults file in the init script
Sometimes it is useful to pass some parameters to NetworkManager when it
starts (e.g. --log-level) instead of editting NetworkManager.conf. Allow
the user add a file with a NETWORKMANAGER_ARGS variable containing such
flags. This is simpler than overriding the whole startup script (e.g. by
means of a rootfs overlay).

Signed-off-by: Carlos Santos <casantos@datacom.ind.br>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Change-Id: Ie8d2d02770cab3f19e42ba3c77dda420f47b3e14