package/wpebackend-fdo: package needs wayland libegl
Now wpebackend-fdo depends only on HAS_LIBEGL but if libegl is not of
wayland type, build breaks due to different EGLNativeWindowType types
depending on backends(fb,x11,wl).

Modify:
'depend on BR2_PACKAGE_HAS_LIBEGL'
to:
'depend on BR2_PACKAGE_HAS_LIBEGL_WAYLAND'
in Config.in to avoid building if there is not a wayland egl backend.
Modify also comment in Config.in when package is not selectable
mentioning the need to have an OpenEGL-capable Wayland backend.

Fixes:
http://autobuild.buildroot.net/results/4f02b91f6ffffd194e09ed18c917b4f678b1a52d/

Signed-off-by: Giulio Benetti <giulio.benetti@micronovasrl.com>
Reviewed-by: Adrian Perez de Castro <aperez@igalia.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
Change-Id: Ica85b031bfe421e4bd3cf1dd360c5d4bdee36ed0