package/wpewebkit: security bump to version 2.26.2
This is a minor release which includes fixes for CVE-2019-8812 and
CVE-2019-8814.

This release also fixes the build with WebDriver disabled, making patch
"0002-WPE-GTK-Build-fails-with-ENABLE_WEBDRIVER-OFF.patch" unneeded
(and therefore removed). There is also a performance improvement for
a regression related to fallback font selection, and a couple of small
fixes. The full release notes are available at:

  https://wpewebkit.org/release/wpewebkit-2.26.2.html

The detailed security advisory can be found at:

  https://wpewebkit.org/security/WSA-2019-0006.html

Signed-off-by: Adrian Perez de Castro <aperez@igalia.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Change-Id: I61a21bfae21ba7357a3f7b2bedd2a799bbc3788b