autotools: do not overwrite first include path
The first include path is special in aclocal. For example it is the path
for the --install option. Also, the first include is treated in a
special way if it doesn't exists. This might be the case if there is the
following construct:

  configure.ac: AC_CONFIG_MACRO_DIR([m4])
  Makefile.am: ACLOCAL_AMFLAGS="-I m4"

If the package doesn't have local macros, the m4/ directory might not
exist. aclocal will then just issue a warning instead of aborting the
execution with a fatal error. See discussion here:
  https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=565663

Don't use the "-I" option in aclocal. Instead use ACLOCAL_PATH to pass
the system-wide include dirs.

As a side effect this should fix the use of $(ACLOCAL) alone. Up until
now, $(ACLOCAL) didn't include the ACLOCAL_HOST_DIR system include path.

autoreconf will pass the "-I" options to every tool it runs, of which
aclocal, which, as seen above, we don't want. So move the argument down
to each individual tool, except for aclocal.

Signed-off-by: Michael Walle <michael@walle.cc>
Signed-off-by: Heiko Thiery <heiko.thiery@gmail.com>
[yann.morin.1998@free.fr: slight rewording of the commit log]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
(cherry picked from commit d255b67972b4b7f27572581fe0c8c8aa03d850c8)

Change-Id: Ia1defc53c6adf1a0e9654b72277659505ab48b90
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>