build: Makefile: change for rockchip platform
1) This adds a defconfig_hook.py to split/merge defconfig while
the %_defconfig or savedefconfig are called. It supports "#include" tags to
load CONFIG from files in configs/rockchip/ directory, to keep a relative
brevity of %_defconfig.

2) modify the /etc/version contents

3) Set output dir to ${TARGET_OUTPUT_DIR} when running make under
buildroot dir.

4) Support reinstalling everything.
(glibc and some packages would be rebuilt)

5) unexport 'UNZIP' environment variables

Having 'UNZIP' in the environment would make the Makefile export it's
'UNZIP' variables to child processes, which would confuse the real unzip
program.

This commit fixes:
Archive:  /usr/bin/unzip
End-of-central-directory signature not found.  Either this file is not
a zipfile, or it constitutes one disk of a multi-part archive.  In the
latter case the central directory and zipfile comment will be found on
the last disk(s) of this archive.

Signed-off-by: Caesar Wang <wxt@rock-chips.com>
Change-Id: I7f4dd15e356308bd0e0a55c3fa9ef55ff1f9dd95